Added support for native input placeholders
The placeholder look-alike doesn't act properly when username/password is saved and the browser automatically fills out the input elements. It appears that the change event is not fired.

Thankfully, most browsers support the placeholder attribute which does exactly what the code intends to do so this change favors the native implementation and falls back to the old implementation.