Improves fix to long email retrieval times
Originally we hard-coded the email tab to limit the number to 500. Some users still want to see all
messages, so now we have a 'Show all' option hidden behind an 'Advanced options' button. It will probably
still time out for users with large numbers of messages, but they've been warned.